agent/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
ddgs>=9.0.0
qrcode[pil]>=7.4.2
pyTelegramBotAPI>=4.14.0
pdfplumber>=0.10.0
python-docx>=1.1.0
